refactor: utilities folder for common services.
