+++ src/common/index.ts
@@ -1,7 +1,5 @@
-export * from './substrate';
-export * from './proxies/date-time';
-export * from './mailer';
-export * from './proxies/process-env';
+export * from './utilities';
+export * from './proxies';
 export * from './polkadot-provider';
 export * from './interceptors';
 export * from './health-indicators';
