+++ src/common/utilities/mailer/models/customer-staking-request-service.model.ts
