+++ src/common/utilities/mailer/templates/customer-staking-request-service.hbs
