+++ src/common/utilities/mailer/templates/lab-register.hbs
