+++ src/common/utilities/reward/index.ts
@@ -1,0 +1,2 @@
+export * from './reward.module';
+export * from './reward.service';
