+++ src/common/utilities/transaction-logging/index.ts
@@ -1,0 +1,2 @@
+export * from './transaction-logging.module';
+export * from './transaction-logging.service';
