+++ src/endpoints/escrow/escrow.controller.ts
@@ -1,5 +1,5 @@
 import { Controller, UseInterceptors } from '@nestjs/common';
-import { SentryInterceptor } from '../../common';
+import { SentryInterceptor } from '../../common/interceptors';
 
 @UseInterceptors(SentryInterceptor)
 @Controller('escrow')
