+++ src/endpoints/substrate-indexed-data/services/service.controller.ts
@@ -1,6 +1,6 @@
 import { Controller, Get, Param, Req, UseInterceptors } from '@nestjs/common';
 import { ApiParam } from '@nestjs/swagger';
-import { SentryInterceptor } from '../../../common';
+import { SentryInterceptor } from '../../../common/interceptors';
 import { ServiceService } from './service.service';
 
 @UseInterceptors(SentryInterceptor)
