+++ src/endpoints/verification/verification.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { DateTimeModule } from '../../common/proxies/date-time/date-time.module';
 import { ProcessEnvModule } from '../../common/proxies/process-env/process-env.module';
-import { RewardModule } from '../../common/reward/reward.module';
+import { RewardModule } from '../../common/utilities/reward/reward.module';
 import { SubstrateModule } from '../../substrate/substrate.module';
 import { VerificationController } from './verification.controller';
 import { VerificationService } from './verification.service';
