+++ src/substrate/orderEvent.ts
@@ -1,14 +1,14 @@
 import { ApiPromise } from '@polkadot/api';
 import { Option } from '@polkadot/types';
 import { EscrowService } from '../endpoints/escrow/escrow.service';
-import { RewardDto } from '../common/reward/dto/reward.dto';
-import { RewardService } from '../common/reward/reward.service';
+import { RewardDto } from '../common/utilities/reward/dto/reward.dto';
+import { RewardService } from '../common/utilities/reward/reward.service';
 import { SubstrateService } from './substrate.service';
 import { Logger } from '@nestjs/common';
-import { TransactionLoggingService } from '../common/transaction-logging/transaction-logging.service';
-import { TransactionLoggingDto } from '../common/transaction-logging/dto/transaction-logging.dto';
+import { TransactionLoggingService } from '../common/utilities/transaction-logging/transaction-logging.service';
+import { TransactionLoggingDto } from '../common/utilities/transaction-logging/dto/transaction-logging.dto';
 import { ethers } from 'ethers'
-import { DebioConversionService } from '../common/debio-conversion/debio-conversion.service';
+import { DebioConversionService } from '../common/utilities/debio-conversion/debio-conversion.service';
 
 export class OrderEventHandler {
   constructor(
