+++ src/substrate/substrate.module.ts
@@ -2,11 +2,11 @@ import { forwardRef, Module } from '@nestjs/common';
 import { SubstrateService } from './substrate.service';
 import { SubstrateController } from './substrate.controller';
 import { EscrowModule } from '../endpoints/escrow/escrow.module';
-import { TransactionLoggingModule } from '../common/transaction-logging/transaction-logging.module';
-import { RewardModule } from '../common/reward/reward.module';
-import { MailModule } from '../common/mailer/mailer.module';
+import { TransactionLoggingModule } from '../common/utilities/transaction-logging/transaction-logging.module';
+import { RewardModule } from '../common/utilities/reward/reward.module';
+import { MailModule } from '../common/utilities/mailer/mailer.module';
 import { LocationModule } from '../endpoints/location/location.module';
-import { DebioConversionModule } from '../common/debio-conversion/debio-conversion.module';
+import { DebioConversionModule } from '../common/utilities/debio-conversion/debio-conversion.module';
 
 @Module({
   imports: [
